print JDL in info
