+++ generators/export-jdl/index.js
@@ -19,7 +19,6 @@
 const util = require('util');
 const chalk = require('chalk');
 const generator = require('yeoman-generator');
-const jhiCore = require('jhipster-core');
 const BaseGenerator = require('../generator-base');
 
 const ExportJDLGenerator = generator.extend({});
@@ -30,7 +29,6 @@ module.exports = ExportJDLGenerator.extend({
     constructor: function (...args) { // eslint-disable-line object-shorthand
         generator.apply(this, args);
         this.baseName = this.config.get('baseName');
-        this.jdl = new jhiCore.JDLObject();
         this.argument('jdlFile', { type: String, required: false, defaults: `${this.baseName}.jh` });
         this.jdlFile = this.options.jdlFile;
     },
@@ -43,15 +41,7 @@ module.exports = ExportJDLGenerator.extend({
 
         parseJson() {
             this.log('Parsing entities from .jhipster dir...');
-            try {
-                const entities = {};
-                this.getExistingEntities().forEach((entity) => { entities[entity.name] = entity.definition; });
-                jhiCore.convertJsonEntitiesToJDL(entities, this.jdl);
-                jhiCore.convertJsonServerOptionsToJDL({ 'generator-jhipster': this.config.getAll() }, this.jdl);
-            } catch (e) {
-                this.log(e.message || e);
-                this.error('\nError while parsing entities to JDL\n');
-            }
+            this.jdl = this.generateJDLFromEntities();
         }
     },
